exercice 3 : passage float en double

+++ exercices/exercice3.cs
@@ -1,8 +1,8 @@
 Console.WriteLine("choisis un nombre");
-float nb1 = float.Parse(Console.ReadLine()?? "0");
+double nb1 = double.Parse(Console.ReadLine()?? "0");
 Console.WriteLine("premier nombre : " + nb1);
 Console.WriteLine("choisis un 2eme nombre");
-float nb2 = float.Parse(Console.ReadLine()?? "0");
+double nb2 = double.Parse(Console.ReadLine()?? "0");
 Console.WriteLine("deuxieme nombre : " + nb2);
 Console.WriteLine("addition : " + (nb1 + nb2));
 Console.WriteLine("soustraction : " + (nb1 - nb2));
